+++ src/Game/Werewolf.hs
@@ -12,9 +12,10 @@ Re-exports all of the public modules under /Game.Werewolf/. These are:
 * "Game.Werewolf.Player"
 * "Game.Werewolf.Response"
 * "Game.Werewolf.Role"
+* "Game.Werewolf.Variant"
 
-N.B., where clashes are found between "Game.Werewolf.Player" and "Game.Werewolf.Role", the
-"Game.Werewolf.Player" functions are preferred.
+N.B., where clashes are found between "Game.Werewolf.Player", "Game.Werewolf.Role" and
+"Game.Werewolf.Variant", the "Game.Werewolf.Player" functions are preferred.
 -}
 
 module Game.Werewolf (
@@ -25,3 +26,4 @@ import Game.Werewolf.Game     as Werewolf
 import Game.Werewolf.Player   as Werewolf
 import Game.Werewolf.Response as Werewolf
 import Game.Werewolf.Role     as Werewolf hiding (activity, name)
+import Game.Werewolf.Variant  as Werewolf hiding (name, tag)